// src/payload-types.ts
export interface Media {
  id: string
  alt: string
  caption?: string | null
  url?: string | null
  filename?: string | null
  mimeType?: string | null
  filesize?: number | null
  width?: number | null
  height?: number | null
  createdAt: string
  updatedAt: string
}

export interface Category {
  id: string
  name: string
  slug: string
  color?: ('terracotta' | 'sage' | 'blue' | 'sand') | null
  createdAt: string
  updatedAt: string
}

export interface Author {
  id: string
  name: string
  title?: string | null
  avatar?: (string | null) | Media
  bio?: string | null
  createdAt: string
  updatedAt: string
}

export interface Product {
  id: string
  name: string
  slug: string
  image?: (string | null) | Media
  price: number
  category?: (string | null) | Category
  description?: string | null
  material?: string | null
  dimensions?: string | null
  weight?: string | null
  colors?:
    | {
        label: string
        hex: string
        id?: string | null
      }[]
    | null
  inStock?: boolean | null
  createdAt: string
  updatedAt: string
}